% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_year.R
\name{add_year}
\alias{add_year}
\title{add_year}
\usage{
add_year(dataframe, yearofdata)
}
\arguments{
\item{dataframe}{the dataframe lacking a year column}

\item{yearofdata}{the year the user wants to add}
}
\value{
the dataframe with the year as its first column
}
\description{
add_year
}
