% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_analysis.R
\name{country_analysis}
\alias{country_analysis}
\title{Country analysis}
\usage{
country_analysis(country)
}
\arguments{
\item{country}{the country you want to analyze}
}
\value{
the graph with the evolution of the main indicators
}
\description{
Country analysis
}
