% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria_analysis.R
\name{criteria_analysis}
\alias{criteria_analysis}
\title{criteria_analysis}
\usage{
criteria_analysis(criteria, top, focusyear)
}
\arguments{
\item{criteria}{the criteria we want to focus on, which should be among the columns of the Happiness dataset}

\item{top}{the number of countries we want to list}

\item{focusyear}{the year we want to focus on}
}
\value{
the top countries for one given criteria
}
\description{
criteria_analysis
}
