% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2015.R
\docType{data}
\name{data2015}
\alias{data2015}
\title{data2015}
\description{
data2015
}
\references{
\url{kaggle.com}{https://www.kaggle.com/unsdsn/world-happiness}
}
\author{
Kaggle \email{vincent@thinkr.fr}
}
\keyword{data}
\keyword{happiness}
\keyword{world}
