% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2017.R
\docType{data}
\name{data2017}
\alias{data2017}
\title{data2017}
\description{
data2017
}
\references{
\url{kaggle.com}{https://www.kaggle.com/unsdsn/world-happiness}
}
\author{
Kaggle \email{alexandre.chouraki@hec.edu}
}
\keyword{data}
\keyword{happiness}
\keyword{world}
