% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Happiness.R
\docType{data}
\name{data_Happiness}
\alias{data_Happiness}
\title{data_Happiness}
\description{
data_Happiness
}
\references{
\url{kaggle.com}{https://www.kaggle.com/unsdsn/world-happiness}
}
\author{
Kaggle \email{alexandre.chouraki@hec.edu}
}
\keyword{data}
\keyword{happiness}
\keyword{world}
