% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2016.R
\docType{data}
\name{data2016}
\alias{data2016}
\title{data2016}
\description{
data2016
}
\references{
\url{kaggle.com}{https://www.kaggle.com/unsdsn/world-happiness}
}
\author{
Kaggle \email{alexandre.chouraki@hec.edu}
}
\keyword{data}
\keyword{happiness}
\keyword{world}
